requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0
selenium==4.25.0
webdriver-manager==4.0.2
extra-streamlit-components==0.1.71